refactor: singleton briefing, best-effort publication, and github client decomposition
